rename copydir to fs
